Bir çok değişiklik ve ekleme yapıldı.
- Parametre ve Data tabloları eklendi.
- Yeni datalar ve parametreler client trafında uygulanıyor.
- Defans / Gemi /  Araştırma daki buglar düzeltildi. Buna göre artık her zaman kaynak hesaplamaları yapılıyor yani gezegende olunmasa bile diğer gezegen içinde hesaplama yapılıyor.
- Filo ve Defanslar düzgün şekilde çalışmıyordu. Buradaki buglar düzeltildi.

+++ WarOfGalaxiesUnity/War Of Galaxies/Assets/PlanetActionController.cs
@@ -329,7 +329,7 @@ public class PlanetActionController : BasePanelController
         float shipSpeed = SLIDER_ShipSpeed.value / 100;
 
         // Uçuş süresi hesaplanacak.
-        int minSpeed = FleetExtends.GetMinSpeedInFleet(ShipsToSend.Select(x => x.UserPlanetShip.ShipId));
+        double minSpeed = FleetExtends.GetMinSpeedInFleet(ShipsToSend.Select(x => x.UserPlanetShip.ShipId));
 
         // Gezegenler arasındaki mesafe.
         double distance = FleetExtends.CalculateDistance(GlobalPlanetController.GPC.CurrentPlanetCordinate, CurrentCordinate);
